Fix AttributeError in base.html when called without `response_headers`
This happens when using a `404.html` which extends DRF's `base.html` and is called by a `Resolver404`.
Since `Resolver404` goes via django's `handler404` and not via DRF's exception handling,
it doesn't get the extra context.

Since the `|items` filter doesn't check for `response_headers` being null (missing) it attempts
to call `<None>.items()`, thus raising an AttributeError.

Originally I solved this in the template by guarding the `for` loop with a `if response_headers`.
However @rpkilby expressed a preference for doing the check inside the `items` filter instead.